app/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
beanie==1.12.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.19.0
yarl==1.8.1
